+++ docs/current_docs/quickstart/index.mdx
@@ -8,7 +8,7 @@ title: "Introduction"
 
 Welcome to Dagger, a programmable tool that lets you replace your software project's artisanal scripts with a modern API and cross-language scripting engine.
 
-Dagger can help you create a consistent, reliable development and testing environment on your local host and ensure that everything your CI can do, your local environment can do too.
+Dagger helps you create a consistent, reliable development and testing environment on your local host and ensure that everything your CI can do, your local environment can do too.
 
 ## Topics
 
@@ -21,18 +21,19 @@ This quickstart introduces you to:
 
 ## Goal
 
-After going through this quickstart, you will have the tools and skills to create a CI pipeline to test, build and release your code, and start replacing your custom CI scripts with reusable Dagger Functions from the Daggerverse or that you create yourself.
+After going through this quickstart, you will have the tools and skills to test, build and release your application using Dagger Functions.
 
 ## Approach
 
-This quickstart is structured as a hands-on workshop, in which you will create and run a CI pipeline for an example application locally. It discusses two approaches:
+This quickstart is structured as a hands-on workshop, in which you will create and run Dagger Functions to test, build and publish an example application.
 
-1. Using the core Dagger API: To begin, you will build a CI pipeline for the example application using only the Dagger API. This is an important first step to learn how Dagger works and to understand key concepts and techniques. It is critical knowledge for when no Daggerverse module exists for your requirements and you need to start from scratch.
-
-1. Using a Daggerverse module: At a later stage, you will refactor the same CI pipeline using a module from the Daggerverse. This alternative approach introduces you to the Daggerverse, a key component of the Dagger ecosystem. It demonstrates the speed and efficiency benefits of reusing existing Dagger modules when constructing your CI pipeline.
+- To begin, you will build a delivery pipeline for the example application using only the Dagger API. This is an important first step to learn how Dagger works and to understand key concepts and techniques.
+- Next, you will simplify the pipeline using a Dagger module from the Daggerverse. This introduces you to the Daggerverse and the reusability benefits it offers.
 
 ## Requirements
 
-This quickstart will take you approximately 40 minutes to complete. You should be familiar with programming in Go, Python or TypeScript, and you should have a basic understanding of [what Dagger is and how it works](../index.mdx).
+This quickstart will take you approximately 20 minutes to complete. You should be familiar with programming in Go, Python or TypeScript.
 
-Before beginning, ensure that you have [Git](https://git-scm.com/downloads) and a container runtime installed on your system. This can be [Docker](https://docs.docker.com/engine/install/), but you can also use Podman, nerdctl, or other Docker-like systems.
+Before beginning, ensure that:
+- you know [what Dagger is](../index.mdx).
+- you have [Git](https://git-scm.com/downloads) and a container runtime installed on your system. This can be [Docker](https://docs.docker.com/engine/install/), but you can also use Podman, nerdctl, or other Docker-like systems.
