+++ docs/current_docs/quickstart/simplify.mdx
@@ -1,18 +1,17 @@
 ---
-slug: /quickstart/refactor
-hide_table_of_contents: true
-title: "Refactor the pipeline using a Daggerverse module"
+slug: /quickstart/simplify
+title: "Simplify with the Daggerverse"
 ---
 import Tabs from "@theme/Tabs";
 import TabItem from "@theme/TabItem";
 
 # Quickstart
 
-## Refactor the pipeline using a Daggerverse module
+## Simplify with the Daggerverse
 
 FIXME USING NEW NODE MODULE
 
-At this point, you have successfully built a CI pipeline for an application using the Dagger API. This is important knowledge to have, so that you understand the basics of using Dagger and can begin creating CI pipelines for your own applications.
+At this point, you have successfully built a pipeline for an application using the Dagger API. This is important knowledge to have, so that you understand the basics of using Dagger and can begin creating  pipelines for your own applications.
 
 That said, Dagger also lets you use Dagger Functions developed by others. This can save you significant time and effort, because it allows you to reuse functionality developed by others without needing to write any code yourself. You can search for useful Dagger Functions in the [Daggerverse](https://daggerverse.dev), Dagger's free and publicly-available index of Dagger modules.
 
@@ -30,7 +29,7 @@ Using the Daggerverse is optional, and does not change how you use Dagger. If yo
 - [Play a game of yamlinvaders](https://daggerverse.dev/mod/github.com/sipsma/daggerverse/yamlinvaders)
 :::
 
-To understand how this works in practice, let's refactor the CI pipeline you just built by using a third-party Dagger module. A good candidate here is the [Node module](https://daggerverse.dev/mod/github.com/dagger/dagger/sdk/typescript/dev/node) from the Daggerverse. This module contains tested, ready-to-use Dagger Functions for building, linting and testing Node.js applications.
+To understand how this works in practice, let's refactor the pipeline you just built by using a third-party Dagger module. A good candidate here is the [Node module](https://daggerverse.dev/mod/github.com/dagger/dagger/sdk/typescript/dev/node) from the Daggerverse. This module contains tested, ready-to-use Dagger Functions for building, linting and testing Node.js applications.
 
 To install the Node module as a dependency, copy and run the `dagger install` command shown on its [Daggerverse page](https://daggerverse.dev/mod/github.com/dagger/dagger/sdk/typescript/dev/node):
