+++ docs/current_docs/quickstart/snippets/daggerize/go/dagger.gen.go
