+++ docs/current_docs/quickstart/snippets/daggerize/go/go.mod
