+++ docs/current_docs/quickstart/snippets/test/go/main.go
@@ -8,25 +8,11 @@ type HelloDagger struct{}
 
 // Returns the result of running unit tests
 func (m *HelloDagger) Test(ctx context.Context, source *Directory) (string, error) {
-	// use the build environment container
-	// run unit tests
+	// get the build environment container
+	// by calling another Dagger Function
 	return m.BuildEnv(source).
+		// call the test runner
 		WithExec([]string{"npm", "run", "test:unit", "run"}).
+		// capture and return the command output
 		Stdout(ctx)
-}
-
-// Returns a container with the build environment
-func (m *HelloDagger) BuildEnv(source *Directory) *Container {
-	// create a Dagger cache volume for dependencies
-	nodeCache := dag.CacheVolume("node")
-	// create the build environment
-	// start from a base node container
-	// add source code
-	// install dependencies
-	return dag.Container().
-		From("node:21-slim").
-		WithDirectory("/src", source.WithoutDirectory("dagger")).
-		WithMountedCache("/src/node_modules", nodeCache).
-		WithWorkdir("/src").
-		WithExec([]string{"npm", "install"})
 }
