+++ docs/current_docs/quickstart/snippets/test/python/__init__.py
@@ -1,0 +1,18 @@
+import dagger
+from dagger import dag, function, object_type
+
+
+@object_type
+class HelloDagger:
+    @function
+    async def test(self, source: dagger.Directory) -> str:
+        """Returns the result of running unit tests"""
+        return await (
+            # get the build environment container
+	          # by calling another Dagger Function
+            self.build_env(source)
+            # call the test runner
+            .with_exec(["npm", "run", "test:unit", "run"])
+            # capture and return the command output
+            .stdout()
+        )
